+++ src/config/keys/mod.rs
@@ -13,20 +13,15 @@ pub use actions::{
     QueueActions,
     SearchActions,
 };
-use actions::{
-    AlbumsActionsFile,
-    ArtistsActionsFile,
-    CommonActionFile,
-    DirectoriesActionsFile,
-    GlobalActionFile,
-    QueueActionsFile,
-};
-use crossterm::event::{KeyCode, KeyModifiers};
+use actions::{CommonActionFile, GlobalActionFile, QueueActionsFile};
 pub use key::Key;
 use serde::{Deserialize, Serialize};
 
 use super::defaults;
-use crate::config::keys::{actions::SaveKind, key::KeySequence};
+use crate::config::keys::{
+    actions::{DuplicateStrategy, RateKind, SaveKind},
+    key::KeySequence,
+};
 
 pub(crate) mod actions;
 pub mod key;
@@ -65,98 +60,115 @@ impl Default for KeyConfigFile {
     fn default() -> Self {
         use GlobalActionFile as G;
         use CommonActionFile as C;
-        use AlbumsActionsFile as Al;
-        use ArtistsActionsFile as Ar;
-        use DirectoriesActionsFile  as D;
-        use KeyCode as K;
-        use KeyModifiers as M;
         #[cfg(debug_assertions)]
         use LogsActionsFile as L;
         use QueueActionsFile as Q;
-        Self {
-            clear: false,
-            global: HashMap::from([
-                (Key { key: K::Char('q'), modifiers: M::NONE  }.into(), G::Quit),
-                (Key { key: K::Char(':'), modifiers: M::NONE  }.into(), G::CommandMode),
-                (Key { key: K::Char('~'), modifiers: M::NONE  }.into(), G::ShowHelp),
-                (Key { key: K::Char('I'), modifiers: M::SHIFT }.into(), G::ShowCurrentSongInfo),
-                (Key { key: K::Char('O'), modifiers: M::SHIFT }.into(), G::ShowOutputs),
-                (Key { key: K::Char('P'), modifiers: M::SHIFT }.into(), G::ShowDecoders),
-                (Key { key: K::Char('>'), modifiers: M::NONE  }.into(), G::NextTrack),
-                (Key { key: K::Char('<'), modifiers: M::NONE  }.into(), G::PreviousTrack),
-                (Key { key: K::Char('s'), modifiers: M::NONE  }.into(), G::Stop),
-                (Key { key: K::Char('z'), modifiers: M::NONE  }.into(), G::ToggleRepeat),
-                (Key { key: K::Char('x'), modifiers: M::NONE  }.into(), G::ToggleRandom),
-                (Key { key: K::Char('c'), modifiers: M::NONE  }.into(), G::ToggleConsume),
-                (Key { key: K::Char('v'), modifiers: M::NONE  }.into(), G::ToggleSingle),
-                (Key { key: K::Char('p'), modifiers: M::NONE  }.into(), G::TogglePause),
-                (Key { key: K::Char('f'), modifiers: M::NONE  }.into(), G::SeekForward),
-                (Key { key: K::Char('b'), modifiers: M::NONE  }.into(), G::SeekBack),
-                (Key { key: K::Char('u'), modifiers: M::NONE  }.into(), G::Update),
-                (Key { key: K::Char('U'), modifiers: M::SHIFT }.into(), G::Rescan),
-                (Key { key: K::Char(','), modifiers: M::NONE  }.into(), G::VolumeDown),
-                (Key { key: K::Char('.'), modifiers: M::NONE  }.into(), G::VolumeUp),
-                (Key { key: K::BackTab,   modifiers: M::SHIFT }.into(), G::PreviousTab),
-                (Key { key: K::Tab,       modifiers: M::NONE  }.into(), G::NextTab),
-                (Key { key: K::Char('R'), modifiers: M::SHIFT }.into(), G::AddRandom),
-                (Key { key: K::Char('1'), modifiers: M::NONE  }.into(), G::SwitchToTab("Queue".to_string())),
-                (Key { key: K::Char('2'), modifiers: M::NONE  }.into(), G::SwitchToTab("Directories".to_string())),
-                (Key { key: K::Char('3'), modifiers: M::NONE  }.into(), G::SwitchToTab("Artists".to_string())),
-                (Key { key: K::Char('4'), modifiers: M::NONE  }.into(), G::SwitchToTab("Album Artists".to_string())),
-                (Key { key: K::Char('5'), modifiers: M::NONE  }.into(), G::SwitchToTab("Albums".to_string())),
-                (Key { key: K::Char('6'), modifiers: M::NONE  }.into(), G::SwitchToTab("Playlists".to_string())),
-                (Key { key: K::Char('7'), modifiers: M::NONE  }.into(), G::SwitchToTab("Search".to_string())),
-            ]),
-            navigation: HashMap::from([
-                (Key { key: K::Char('k'), modifiers: M::NONE    }.into(), C::Up),
-                (Key { key: K::Char('j'), modifiers: M::NONE    }.into(), C::Down),
-                (Key { key: K::Char('l'), modifiers: M::NONE    }.into(), C::Right),
-                (Key { key: K::Left,      modifiers: M::NONE    }.into(), C::Left),
-                (Key { key: K::Up,        modifiers: M::NONE    }.into(), C::Up),
-                (Key { key: K::Down,      modifiers: M::NONE    }.into(), C::Down),
-                (Key { key: K::Right,     modifiers: M::NONE    }.into(), C::Right),
-                (Key { key: K::Char('h'), modifiers: M::NONE    }.into(), C::Left),
-                (Key { key: K::Char('k'), modifiers: M::CONTROL }.into(), C::PaneUp),
-                (Key { key: K::Char('j'), modifiers: M::CONTROL }.into(), C::PaneDown),
-                (Key { key: K::Char('l'), modifiers: M::CONTROL }.into(), C::PaneRight),
-                (Key { key: K::Char('h'), modifiers: M::CONTROL }.into(), C::PaneLeft),
-                (Key { key: K::Char('K'), modifiers: M::SHIFT   }.into(), C::MoveUp),
-                (Key { key: K::Char('J'), modifiers: M::SHIFT   }.into(), C::MoveDown),
-                (Key { key: K::Char('d'), modifiers: M::CONTROL }.into(), C::DownHalf),
-                (Key { key: K::Char('u'), modifiers: M::CONTROL }.into(), C::UpHalf),
-                (Key { key: K::Char('G'), modifiers: M::SHIFT   }.into(), C::Bottom),
-                (Key { key: K::Char('g'), modifiers: M::NONE    }.into(), C::Top),
-                (Key { key: K::Char('/'), modifiers: M::NONE    }.into(), C::EnterSearch),
-                (Key { key: K::Char('n'), modifiers: M::NONE    }.into(), C::NextResult),
-                (Key { key: K::Char('N'), modifiers: M::SHIFT   }.into(), C::PreviousResult),
-                (Key { key: K::Char(' '), modifiers: M::NONE    }.into(), C::Select),
-                (Key { key: K::Char(' '), modifiers: M::CONTROL }.into(), C::InvertSelection),
-                (Key { key: K::Char('a'), modifiers: M::NONE    }.into(), C::Add),
-                (Key { key: K::Char('A'), modifiers: M::SHIFT   }.into(), C::AddAll),
-                (Key { key: K::Char('D'), modifiers: M::SHIFT   }.into(), C::Delete),
-                (Key { key: K::Char('r'), modifiers: M::NONE    }.into(), C::Rename),
-                (Key { key: K::Char('c'), modifiers: M::CONTROL }.into(), C::Close),
-                (Key { key: K::Esc,       modifiers: M::NONE    }.into(), C::Close),
-                (Key { key: K::Enter,     modifiers: M::NONE    }.into(), C::Confirm),
-                (Key { key: K::Char('i'), modifiers: M::NONE    }.into(), C::FocusInput),
-                (Key { key: K::Char('B'), modifiers: M::SHIFT   }.into(), C::ShowInfo),
-                (Key { key: K::Char('z'), modifiers: M::CONTROL }.into(), C::ContextMenu {}),
-                (Key { key: K::Char('s'), modifiers: M::CONTROL }.into(), C::Save { kind: SaveKind::default() }),
-            ]),
-            #[cfg(debug_assertions)]
-            logs: HashMap::from([
-                (Key { key: K::Char('D'), modifiers: M::SHIFT   }.into(), L::Clear),
-                (Key { key: K::Char('S'), modifiers: M::SHIFT   }.into(), L::ToggleScroll),
-            ]),
-            queue: HashMap::from([
-                (Key { key: K::Char('d'), modifiers: M::NONE    }.into(), Q::Delete),
-                (Key { key: K::Char('D'), modifiers: M::SHIFT   }.into(), Q::DeleteAll),
-                (Key { key: K::Enter,     modifiers: M::NONE    }.into(), Q::Play),
-                (Key { key: K::Char('a'), modifiers: M::NONE    }.into(), Q::AddToPlaylist),
-                (Key { key: K::Char('C'), modifiers: M::SHIFT   }.into(), Q::JumpToCurrent),
-                (Key { key: K::Char('X'), modifiers: M::SHIFT   }.into(), Q::Shuffle),
-            ]),
-        }
+
+        let s = || KeySequence::new();
+
+        let global = HashMap::from([
+            (s().char('q'),                       G::Quit),
+            (s().char('?'),                       G::ShowHelp),
+            (s().char(':'),                       G::CommandMode),
+            (s().char('o').char('I'),             G::ShowCurrentSongInfo),
+            (s().char('o').char('o'),             G::ShowOutputs),
+            (s().char('o').char('p'),             G::ShowDecoders),
+            (s().char('o').char('d'),             G::ShowDownloads),
+            (s().char('o').char('P'),             G::Partition { name: None, autocreate: false }),
+            (s().char('z'),                       G::ToggleRepeat),
+            (s().char('x'),                       G::ToggleRandom),
+            (s().char('c'),                       G::ToggleConsume),
+            (s().char('v'),                       G::ToggleSingle),
+            (s().char('p'),                       G::TogglePause),
+            (s().char('s'),                       G::Stop),
+            (s().char('>'),                       G::NextTrack),
+            (s().char('<'),                       G::PreviousTrack),
+            (s().char('f'),                       G::SeekForward),
+            (s().char('b'),                       G::SeekBack),
+            (s().char(','),                       G::VolumeDown),
+            (s().char('.'),                       G::VolumeUp),
+            (s().tab(),                           G::NextTab),
+            (s().char('g').char('t'),             G::NextTab),
+            (s().tab().shift(),                   G::PreviousTab),
+            (s().char('g').char('T'),             G::PreviousTab),
+            (s().char('1'),                       G::SwitchToTab("Queue".to_string())),
+            (s().char('2'),                       G::SwitchToTab("Directories".to_string())),
+            (s().char('3'),                       G::SwitchToTab("Artists".to_string())),
+            (s().char('4'),                       G::SwitchToTab("Album Artists".to_string())),
+            (s().char('5'),                       G::SwitchToTab("Albums".to_string())),
+            (s().char('6'),                       G::SwitchToTab("Playlists".to_string())),
+            (s().char('7'),                       G::SwitchToTab("Search".to_string())),
+            (s().char('u').ctrl(),                G::Update),
+            (s().char('U').ctrl(),                G::Rescan),
+            (s().char('R'),                       G::AddRandom),
+        ]);
+
+        let navigation = HashMap::from([
+            (s().esc(),                           C::Close),
+            (s().char('c').ctrl(),                C::Close),
+            (s().cr(),                            C::Confirm),
+            (s().char('k'),                       C::Up),
+            (s().up(),                            C::Up),
+            (s().char('j'),                       C::Down),
+            (s().down(),                          C::Down),
+            (s().char('h'),                       C::Left),
+            (s().left(),                          C::Left),
+            (s().char('l'),                       C::Right),
+            (s().right(),                         C::Right),
+            (s().char('w').ctrl().char('k'),      C::PaneUp),
+            (s().up().ctrl(),                     C::PaneUp),
+            (s().char('w').ctrl().char('j'),      C::PaneDown),
+            (s().down().ctrl(),                   C::PaneDown),
+            (s().char('w').ctrl().char('h'),      C::PaneLeft),
+            (s().left().ctrl(),                   C::PaneLeft),
+            (s().char('w').ctrl().char('l'),      C::PaneRight),
+            (s().right().ctrl(),                  C::PaneRight),
+            (s().char('K'),                       C::MoveUp),
+            (s().char('J'),                       C::MoveDown),
+            (s().char('u').ctrl(),                C::UpHalf),
+            (s().char('d').ctrl(),                C::DownHalf),
+            (s().char('b').ctrl(),                C::PageUp),
+            (s().page_up(),                       C::PageUp),
+            (s().char('f').ctrl(),                C::PageDown),
+            (s().page_down(),                     C::PageDown),
+            (s().char('g').char('g'),             C::Top),
+            (s().char('G'),                       C::Bottom),
+            (s().char('/'),                       C::EnterSearch),
+            (s().char('n'),                       C::NextResult),
+            (s().char('N'),                       C::PreviousResult),
+            (s().char(' '),                       C::Select),
+            (s().char(' ').ctrl(),                C::InvertSelection),
+            (s().char('a'),                       C::Add),
+            (s().char('A'),                       C::AddAll),
+            (s().char('D'),                       C::Delete),
+            (s().char('r').ctrl(),                C::Rename),
+            (s().char('i'),                       C::FocusInput),
+            (s().char('o').char('i'),             C::ShowInfo),
+            (s().char('z').ctrl(),                C::ContextMenu {}),
+            (s().char('s').ctrl().char('s'),      C::Save { kind: SaveKind::Modal { all: false, duplicates_strategy: DuplicateStrategy::Ask } }),
+            (s().char('s').ctrl().char('a'),      C::Save { kind: SaveKind::Modal { all: true, duplicates_strategy: DuplicateStrategy::Ask } }),
+            (s().char('r'),                       C::Rate { kind: RateKind::default(), current: false, min_rating: 0, max_rating: 10 }),
+        ]);
+
+        let queue = HashMap::from([
+            (s().char('d'),                       Q::Delete),
+            (s().char('D'),                       Q::DeleteAll),
+            (s().cr(),                            Q::Play),
+            (s().char('C'),                       Q::JumpToCurrent),
+            (s().char('X'),                       Q::Shuffle),
+        ]);
+
+        #[cfg(debug_assertions)]
+        let logs = HashMap::from([
+            (s().char('D'),                       L::Clear),
+            (s().char('S'),                       L::ToggleScroll),
+        ]);
+
+        #[cfg(not(debug_assertions))]
+        return KeyConfigFile { clear: false, global, navigation, queue };
+
+        #[cfg(debug_assertions)]
+        return KeyConfigFile { clear: false, global, navigation, queue, logs };
     }
 }
 
@@ -277,7 +289,7 @@ mod tests {
             #[cfg(debug_assertions)]
             logs: HashMap::from([(Key { key: KeyCode::Char('a'), modifiers: KeyModifiers::CONTROL, }.into(), LogsActionsFile::Clear)]),
             queue: HashMap::from([(Key { key: KeyCode::Char('a'), modifiers: KeyModifiers::CONTROL, }.into(), QueueActionsFile::Play),
-                                  (Key { key: KeyCode::Char('b'), modifiers: KeyModifiers::SHIFT, }.into(), QueueActionsFile::Save)]),
+                                  (Key { key: KeyCode::Char('b'), modifiers: KeyModifiers::SHIFT, }.into(), QueueActionsFile::JumpToCurrent)]),
             navigation: HashMap::from([
                 (Key { key: KeyCode::Char('a'), modifiers: KeyModifiers::CONTROL, }.into(), CommonActionFile::Up),
                 (Key { key: KeyCode::Char('b'), modifiers: KeyModifiers::SHIFT }.into(), CommonActionFile::Up)
@@ -288,7 +300,7 @@ mod tests {
             #[cfg(debug_assertions)]
             logs: HashMap::from([(Key { key: KeyCode::Char('a'), modifiers: KeyModifiers::CONTROL, }.into(), LogsActions::Clear)]),
             queue: HashMap::from([(Key { key: KeyCode::Char('a'), modifiers: KeyModifiers::CONTROL, }.into(), QueueActions::Play),
-                                  (Key { key: KeyCode::Char('b'), modifiers: KeyModifiers::SHIFT, }.into(), QueueActions::Save)]),
+                                  (Key { key: KeyCode::Char('b'), modifiers: KeyModifiers::SHIFT, }.into(), QueueActions::JumpToCurrent)]),
             albums: HashMap::from([]),
             artists: HashMap::from([]),
             directories: HashMap::from([]),
@@ -313,7 +325,7 @@ mod tests {
                 (Key { key: KeyCode::Char(' '), modifiers: KeyModifiers::NONE }.into(), GlobalActionFile::TogglePause),
             ]),
             queue: HashMap::from([(Key { key: KeyCode::Char('a'), modifiers: KeyModifiers::CONTROL, }.into(), QueueActionsFile::Play),
-                                  (Key { key: KeyCode::Char('b'), modifiers: KeyModifiers::SHIFT, }.into(), QueueActionsFile::Save)]),
+                                  (Key { key: KeyCode::Char('b'), modifiers: KeyModifiers::SHIFT, }.into(), QueueActionsFile::JumpToCurrent)]),
             navigation: HashMap::from([
                 (Key { key: KeyCode::Char('a'), modifiers: KeyModifiers::CONTROL, }.into(), CommonActionFile::Up),
                 (Key { key: KeyCode::Char('b'), modifiers: KeyModifiers::SHIFT }.into(), CommonActionFile::Up),
@@ -325,7 +337,7 @@ mod tests {
         let mut default: KeyConfig = KeyConfig::default();
         default.global.insert(Key { key: KeyCode::Char('a'), modifiers: KeyModifiers::CONTROL, }.into(), GlobalAction::Quit);
         default.queue.insert(Key { key: KeyCode::Char('a'), modifiers: KeyModifiers::CONTROL, }.into(), QueueActions::Play);
-        default.queue.insert(Key { key: KeyCode::Char('b'), modifiers: KeyModifiers::SHIFT, }.into(), QueueActions::Save);
+        default.queue.insert(Key { key: KeyCode::Char('b'), modifiers: KeyModifiers::SHIFT, }.into(), QueueActions::JumpToCurrent);
         default.navigation.insert(Key { key: KeyCode::Char('a'), modifiers: KeyModifiers::CONTROL }.into(), CommonAction::Up);
         default.navigation.insert(Key { key: KeyCode::Char('b'), modifiers: KeyModifiers::SHIFT }.into(), CommonAction::Up);
         #[cfg(debug_assertions)]